+++ changelog.md
@@ -3,13 +3,15 @@
 ### breaking changes
 * Change `PyCFunction` definition `?*const fn ([*c]c.PyObject, [*c]c.PyObject) callconv(.c) ?*c.PyObject` ->
 `*const fn (*c.PyObject, *c.PyObject) callconv(.c) ?*c.PyObject`, PyMethodDef remain unchanged.
+* Change `Err.NoneError([*:0]const u8)` into `Err.noneError([*:0]const u8)`
 
 ### deprecations
 * add `None()`, deprecate `Py_None()`
 
 ### new stuffs
-* add `wrapPyCFunctionDefault`
+* add `wrapPyCFunctionDefault`.
 * add `fromObjectExact` and `isDictExact` in DictObject.
+* add `Err.customError([*:0]const u8)` and CustomError.
 
 ### changes
 * change `@cimport` into `b.addTranslateC` and `c` module, make `py` module able to use debug build.